Alexa Skills/Family Journal/backend/requirements.txt
flask
flask_restful
flask-sqlalchemy
sqlalchemy-cockroachdb
google-cloud-language==1.2.0
scipy
